// src/client/models/FeedMaterialTypeEnum.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * - NEW_MATERIAL_PACK: 新物料
 * - REUSED_MATERIAL_PACK: 上次未用完這次使用的物料
 * - IMPORTED_MATERIAL_PACK: 未標註的物料(第一次上傳的物料)
 * - INSPECTION_MATERIAL_PACK: 巡檢物料
 */
export enum FeedMaterialTypeEnum {
    NEW_MATERIAL_PACK = 'NEW_MATERIAL_PACK',
    REUSED_MATERIAL_PACK = 'REUSED_MATERIAL_PACK',
    IMPORTED_MATERIAL_PACK = 'IMPORTED_MATERIAL_PACK',
    INSPECTION_MATERIAL_PACK = 'INSPECTION_MATERIAL_PACK',
}